predicts the verb (i think?) and gives it icons
